Move validator to the components folder
