+++ lib/problems/components/action.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'problems/common/validator'
+require_relative 'validator'
 
 module Problems
   # Abstraction of an executable action.
