+++ test/components/entity_test.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-require 'problems/common/validator'
 require 'problems/components/entity'
+require 'problems/components/validator'
 
 class EntityTest < Minitest::Test
   class TestEntity < Problems::Entity
